pyproject.toml
[project]
name = "video-analyzer"
version = "0.1.0"
description = "Video upload app with Whisper transcription and Gemini-based corner segmentation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "sqlalchemy>=2.0.32",
  "pydantic-settings>=2.4.0",
  "openai-whisper>=20240930",
  "google-generativeai>=0.7.2",
  "ffmpeg-python>=0.2.0",
  "jinja2>=3.1.4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "httpx>=0.27.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
